Finalizando as validações do endereço

+++ app/Models/Address.php
@@ -4,8 +4,10 @@ namespace App\Models;
 
 
 use Illuminate\Database\Eloquent\Model;
+use App\Models\Validations;
+use App\Validations\basicValidations;
 
-class Address extends Model
+class Address extends Model implements Validations
 {
     protected $table='address';
     public $timestamps = false;
@@ -20,6 +22,44 @@ class Address extends Model
         'Complement',
         'Reference'
     ];
+
+
+    public function validateAtribute($isEmpty,$type,$minimalSize,$maximumSize,$value){   
+        $validationsResult=false;
+        $a=basicValidations::isEmpty($value)!= $isEmpty ? $validationsResult=false : $validationsResult = true;
+        $a=basicValidations::checkType($value,$type)==false ? $validationsResult=false : $validationsResult = true;
+
+        
+        if($type=="string"){
+            $a=basicValidations::minimalStringSize($value,$minimalSize)==false ? $validationsResult=false : $validationsResult = true;
+            $a=basicValidations::maximumStringSize($value,$maximumSize)==false ? $validationsResult=false : $validationsResult = true;
+        }
+        else{
+            $a=basicValidations::minimalNumericSize($value,$minimalSize)==false ? $validationsResult=false : $validationsResult = true;
+            $a=basicValidations::maximumNumericSize($value,$maximumSize)==false ? $validationsResult=false : $validationsResult = true;
+        }
+
+        return $validationsResult;
+    }
+
+
+    public function Validate(){
+         if(($this->validateAtribute(false,"string",5,100,$this->Street))==false||
+            ($this->validateAtribute(false,"string",8,8,$this->ZipCode))==false||
+            ($this->validateAtribute(false,"string",0,10,$this->Number))==false||
+            ($this->validateAtribute(false,"string",5,50,$this->City))==false||
+            ($this->validateAtribute(false,"string",2,50,$this->State))==false||
+            ($this->validateAtribute(true,"string",0,200,$this->Complement))==false||
+            ($this->validateAtribute(true,"string",0,100,$this->Reference)) ==false
+        ){
+                return false;
+        }
+        else{
+                return true;
+        }
+    }
+
+
     public function setAllAtributs($Street="",$Number="",$ZipCode="",$City="",$State="",$Complement="",$Reference=""){
         $this->Street=$Street;
         $this->ZipCode=$ZipCode;
@@ -56,13 +96,23 @@ class Address extends Model
         ];
     }
     public function saveAddress(){
-        return $this->save();
+
+        if($this->Validate()==true) {return $this->save();} else { return "   Algum dos campos não corresponde as especificações";}
     }
-    public function updateAddress(){
-        return $this->save();
+    public function updateAddress($idAddress){
+        if(empty($idAddress)==true){
+            $this->idAddress=$idAddress;
+            if($this->Validate()==true) {return $this->save();} else { return "Algum dos campos não corresponde as especificações";}
+        }else{ return "Parece que houve um erro, tente novamente "; }
     }
-    public function deleteAddress(){
-        $this->delete();
+    public function deleteAddress($idAddress){
+        if(empty($idAddress)==true){
+            $this->getById($idAddress);
+            $this->delete();
+        }
+        else{
+            return "Parece que houve um erro, tente novamente";
+        }
     }
     public function getById($idAddress){
         $consultedAddress=$this->find($idAddress);
